refactor: split git and google auth on 2 files
