+++ src/pages/signin/index.tsx
@@ -1,4 +1,4 @@
-import SignInWithSocialMedia from '@/src/features/auth/signInWithSocialMedia/signInWithSocialMedia'
+import SignInWithSocialMedia from '@/src/features/auth/signInWithSocialMedia/GitAuth'
 import { Inter } from 'next/font/google'
 import Head from 'next/head'
 import Link from 'next/link'
